Move admin pages to blueprint
Still need to move helper functions out of `modules/admin/routes.py`, since the admin routes
had helpers defined inside the route scope and not in a separate file.
